add "move package" context option
